+++ src/main/java/nextstep/auth/AuthConfig.java
@@ -6,14 +6,14 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import nextstep.auth.authentication.BasicAuthenticationFilter2;
+import nextstep.auth.authentication.BasicAuthenticationFilter;
 import nextstep.auth.authentication.BearerTokenAuthenticationFilter;
-import nextstep.auth.authentication.UsernamePasswordAuthenticationFilter2;
+import nextstep.auth.authentication.TokenAuthenticationInterceptor;
+import nextstep.auth.authentication.UsernamePasswordAuthenticationFilter;
 import nextstep.auth.authorization.AuthenticationPrincipalArgumentResolver;
 import nextstep.auth.context.SecurityContextPersistenceFilter;
 import nextstep.auth.service.CustomUserDetails;
 import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenAuthenticationInterceptor2;
 
 @Configuration
 public class AuthConfig implements WebMvcConfigurer {
@@ -28,11 +28,11 @@ public class AuthConfig implements WebMvcConfigurer {
 	@Override
 	public void addInterceptors(InterceptorRegistry registry) {
 		registry.addInterceptor(new SecurityContextPersistenceFilter());
-		registry.addInterceptor(new UsernamePasswordAuthenticationFilter2(customUserDetails))
+		registry.addInterceptor(new UsernamePasswordAuthenticationFilter(customUserDetails))
 			.addPathPatterns("/login/form");
-		registry.addInterceptor(new TokenAuthenticationInterceptor2(customUserDetails, jwtTokenProvider))
+		registry.addInterceptor(new TokenAuthenticationInterceptor(customUserDetails, jwtTokenProvider))
 			.addPathPatterns("/login/token");
-		registry.addInterceptor(new BasicAuthenticationFilter2(customUserDetails));
+		registry.addInterceptor(new BasicAuthenticationFilter(customUserDetails));
 		registry.addInterceptor(new BearerTokenAuthenticationFilter(jwtTokenProvider));
 	}
