+++ src/main/java/nextstep/auth/authentication/UsernamePasswordAuthenticationFilter.java
@@ -5,50 +5,35 @@ import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.springframework.web.servlet.HandlerInterceptor;
-
 import nextstep.auth.context.Authentication;
 import nextstep.auth.context.SecurityContextHolder;
-import nextstep.member.application.LoginMemberService;
-import nextstep.member.domain.LoginMember;
+import nextstep.auth.service.CustomUserDetails;
 
-public class UsernamePasswordAuthenticationFilter implements HandlerInterceptor {
+public class UsernamePasswordAuthenticationFilter extends AuthenticationInterceptor {
 	private static final String PRINCIPAL_NAME = "username";
 	private static final String CREDENTIAL_NAME = "password";
-	
-	private LoginMemberService loginMemberService;
 
-	public UsernamePasswordAuthenticationFilter(LoginMemberService loginMemberService) {
-		this.loginMemberService = loginMemberService;
+	public UsernamePasswordAuthenticationFilter(CustomUserDetails customUserDetails) {
+		super(customUserDetails);
 	}
 
 	@Override
-	public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+	public AuthenticationToken convert(HttpServletRequest request) {
+		Map<String, String[]> parameterMap = request.getParameterMap();
+		String userName = parameterMap.get(PRINCIPAL_NAME)[0];
+		String password = parameterMap.get(CREDENTIAL_NAME)[0];
 
+		return new AuthenticationToken(userName, password);
+	}
+
+	@Override
+	public boolean afterAuthenticate(Authentication authentication, HttpServletResponse response) {
 		try {
-			Map<String, String[]> parameterMap = request.getParameterMap();
-			String userName = parameterMap.get(PRINCIPAL_NAME)[0];
-			String password = parameterMap.get(CREDENTIAL_NAME)[0];
-
-			AuthenticationToken token = new AuthenticationToken(userName, password);
-
-			LoginMember loginMember = loginMemberService.loadUserByUsername(token.getPrincipal());
-
-			if (loginMember == null) {
-				throw new AuthenticationException();
-			}
-
-			if (!loginMember.checkPassword(token.getCredentials())) {
-				throw new AuthenticationException();
-			}
-
-			Authentication authentication = new Authentication(loginMember.getEmail(), loginMember.getAuthorities());
 			SecurityContextHolder.getContext().setAuthentication(authentication);
-
 			return false;
 		} catch (Exception e) {
 			return true;
 		}
-
 	}
+
 }
